Move base db config to user.
