+++ service/src/main/java/bio/terra/user/app/configuration/UserDatabaseConfiguration.java
@@ -1,5 +1,13 @@
 package bio.terra.user.app.configuration;
 
+import java.util.Properties;
+import org.apache.commons.dbcp2.ConnectionFactory;
+import org.apache.commons.dbcp2.DriverManagerConnectionFactory;
+import org.apache.commons.dbcp2.PoolableConnection;
+import org.apache.commons.dbcp2.PoolableConnectionFactory;
+import org.apache.commons.dbcp2.PoolingDataSource;
+import org.apache.commons.pool2.ObjectPool;
+import org.apache.commons.pool2.impl.GenericObjectPool;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -10,7 +18,7 @@ import org.springframework.transaction.annotation.EnableTransactionManagement;
 @Configuration
 @EnableTransactionManagement
 @ConfigurationProperties(prefix = "user.user-database")
-public class UserDatabaseConfiguration extends BaseDatabaseConfiguration {
+public class UserDatabaseConfiguration {
   // These properties control code in the StartupInitializer. We would not use these in production,
   // but they are handy to set for development and testing. There are only three interesting states:
   // 1. initialize is true; upgrade is irrelevant - initialize and recreate an empty database
@@ -20,6 +28,65 @@ public class UserDatabaseConfiguration extends BaseDatabaseConfiguration {
   private boolean initializeOnStart;
   /** If true, primary database will have changesets applied */
   private boolean upgradeOnStart;
+
+  private String uri;
+  private String username;
+  private String password;
+
+  private PoolingDataSource<PoolableConnection> dataSource;
+
+  public String getUri() {
+    return uri;
+  }
+
+  public String getUsername() {
+    return username;
+  }
+
+  public String getPassword() {
+    return password;
+  }
+
+  // NOTE: even though the setters appear unused, the Spring infrastructure uses them to populate
+  // the properties.
+  public void setUri(String uri) {
+    this.uri = uri;
+  }
+
+  public void setUsername(String username) {
+    this.username = username;
+  }
+
+  public void setPassword(String password) {
+    this.password = password;
+  }
+
+  // Main use of the configuration is this pooling data source object.
+  public PoolingDataSource<PoolableConnection> getDataSource() {
+    // Lazy allocation of the data source
+    if (dataSource == null) {
+      configureDataSource();
+    }
+    return dataSource;
+  }
+
+  private void configureDataSource() {
+    Properties props = new Properties();
+    props.setProperty("user", getUsername());
+    props.setProperty("password", getPassword());
+
+    ConnectionFactory connectionFactory = new DriverManagerConnectionFactory(getUri(), props);
+
+    PoolableConnectionFactory poolableConnectionFactory =
+        new PoolableConnectionFactory(connectionFactory, null);
+
+    ObjectPool<PoolableConnection> connectionPool =
+        new GenericObjectPool<>(poolableConnectionFactory);
+
+    poolableConnectionFactory.setPool(connectionPool);
+
+    dataSource = new PoolingDataSource<>(connectionPool);
+  }
 
   public boolean isInitializeOnStart() {
     return initializeOnStart;
